packages/shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities and types for Tiger MCP system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "typing-extensions>=4.0.0",
    "cryptography>=41.0.0",
    "bcrypt>=4.0.0",
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.8.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.42",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.16.4",
    "httpx>=0.28.1",
    "redis>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.28.1",
    "fakeredis>=2.17.0",
    "pytest-redis>=3.0.0",
]
